+++ tests/test_feature_append_concat.py
@@ -98,8 +98,8 @@ def test_concat(test_df, reverse, iterable):
     assert test_df.scenario == ["scen_a", "scen_b"]
     assert other.scenario == ["scen_c"]
 
-    # assert that merging of meta works as expected (reorder columns)
-    pdt.assert_frame_equal(result.meta[EXP_META.columns], EXP_META)
+    # assert that merging of meta works as expected
+    pdt.assert_frame_equal(result.meta[EXP_META.columns], EXP_META, check_like=True)
 
     # assert that appending data works as expected
     ts = result.timeseries()
@@ -188,7 +188,31 @@ def test_concat_all_pd_dataframe(test_df):
     npt.assert_array_equal(ts.iloc[2].values, ts.iloc[3].values)
 
 
-def test_append(test_df):
+@pytest.mark.parametrize("inplace", (True, False))
+def test_append_data_not_sorted(test_pd_df, inplace):
+    """Appending timeseries data does not sort"""
+
+    columns = IAMC_IDX + list(test_pd_df.columns[[6, 5]])
+    unsorted_data = test_pd_df.iloc[[2, 1]][columns]
+    df = IamDataFrame(unsorted_data)
+
+    if inplace:
+        obs = df.copy()
+        obs.append(test_pd_df.iloc[[0]], inplace=True)
+    else:
+        obs = df.append(test_pd_df.iloc[[0]])
+        # assert that original object was not modified
+        assert len(df._data) == 4
+
+    # `data` is not sorted, only applies to pandas >= 2.2
+    # TODO remove this if-statement when dropping support for pandas < 2.2
+    if pd.__version__ >= "2.2":
+        assert list(obs.data.scenario.unique()) == ["scen_b", "scen_a"]
+        assert list(obs.data.year.unique()) == [2010, 2005]
+        assert not obs._data.index.is_monotonic_increasing
+
+
+def test_append_meta(test_df):
     other = test_df.filter(scenario="scen_b").rename({"scenario": {"scen_b": "scen_c"}})
 
     test_df.set_meta([0, 1], name="col1")