+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
-# Next Release
+# Next release
 
+The next release must be bumped to v3.0.0.
+
+- [#893](https://github.com/IAMconsortium/pyam/pull/893) No sorting of timeseries data on initialization or append
 - [#879](https://github.com/IAMconsortium/pyam/pull/879) Add `read_netcdf()` function
 
 # Release v2.3.0